+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/CreateWalletGuidelines.test.tsx
@@ -2,16 +2,17 @@ import { render } from '@testing-library/react-native'
 import { CreateWalletGuidelines } from './CreateWalletGuidelines'
 
 jest.mock('@shared-contexts/ThemeProvider')
-describe('create wallet guidelines', () => {
+describe('create wallet guidelines v2', () => {
   it('should match snapshot', () => {
     const navigation: any = {
       navigate: jest.fn()
     }
     const route: any = {}
-    const rendered = render(<CreateWalletGuidelines
-      navigation={navigation}
-      route={route}
-                            />)
+    const rendered = render(
+      <CreateWalletGuidelines
+        navigation={navigation}
+        route={route}
+      />)
     expect(rendered.toJSON()).toMatchSnapshot()
   })
 })
