+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/PinConfirmation.tsx
@@ -1,38 +1,36 @@
 import { StackScreenProps } from '@react-navigation/stack'
 import { useState } from 'react'
-import { initJellyfishWallet, MnemonicEncrypted } from '@api/wallet'
-import { MnemonicStorage } from '@api/wallet/mnemonic_storage'
 import { View } from '@components/index'
-import { CREATE_STEPS, CreateWalletStepIndicator, RESTORE_STEPS } from '@components/CreateWalletStepIndicator'
-import { PinTextInput } from '@components/PinTextInput'
-import { ThemedActivityIndicator, ThemedScrollView, ThemedText } from '@components/themed'
+import { ThemedActivityIndicatorV2, ThemedScrollViewV2, ThemedTextV2 } from '@components/themed'
 import { useNetworkContext } from '@shared-contexts/NetworkContext'
-import { useWalletPersistenceContext, WalletPersistenceDataI } from '@shared-contexts/WalletPersistenceContext'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import { WalletParamList } from '@screens/WalletNavigator/WalletNavigator'
-import { useLogger } from '@shared-contexts/NativeLoggingProvider'
-import { useWhaleApiClient } from '@shared-contexts/WhaleContext'
-import { WalletAddressIndexPersistence } from '@api/wallet/address_index'
+import { CREATE_STEPS, RESTORE_STEPS, CreateWalletStepIndicatorV2 } from '@components/CreateWalletStepIndicatorV2'
+import { PinTextInputV2 } from '@components/PinTextInputV2'
+import { NavigationProp, useNavigation } from '@react-navigation/native'
+import { WalletPersistenceDataI } from '@shared-contexts/WalletPersistenceContext'
 import { EncryptedProviderData } from '@defichain/jellyfish-wallet-encrypted'
-import { MAX_ALLOWED_ADDRESSES } from '@shared-contexts/WalletContext'
+import { MnemonicEncrypted } from '@api/wallet'
+import { MnemonicStorage } from '@api/wallet/mnemonic_storage'
+import { useLogger } from '@shared-contexts/NativeLoggingProvider'
 
 type Props = StackScreenProps<WalletParamList, 'PinConfirmation'>
 
 export function PinConfirmation ({ route }: Props): JSX.Element {
+  const navigation = useNavigation<NavigationProp<WalletParamList>>()
   const logger = useLogger()
+  const [isComplete, setIsComplete] = useState<boolean>(false) // To complete the last stepper node when pin is verified.
   const { network } = useNetworkContext()
-  const { setWallet } = useWalletPersistenceContext()
-  const client = useWhaleApiClient()
   const {
     pin,
     words,
     type
   } = route.params
   const [newPin, setNewPin] = useState('')
-
   const [invalid, setInvalid] = useState<boolean>(false)
   const [spinnerMessage, setSpinnerMessage] = useState<string>()
+  const isCreateWallet = type === 'create'
 
   function verifyPin (input: string): void {
     if (input.length !== pin.length) {
@@ -44,103 +42,109 @@ export function PinConfirmation ({ route }: Props): JSX.Element {
       return
     } else {
       setInvalid(false)
+      setIsComplete(true)
     }
 
-    const copy = { words, network, pin }
+    const copy = { words, network, pin, isWalletRestored: type === 'restore' }
+
     setSpinnerMessage(translate('screens/PinConfirmation', 'It may take a few seconds to securely encrypt your wallet...'))
     setTimeout(() => {
       MnemonicEncrypted.toData(copy.words, copy.network, copy.pin)
         .then(async encrypted => {
           await MnemonicStorage.set(words, pin)
-          if (type === 'restore') {
-            await discoverWalletAddresses(encrypted)
-          }
-          await setWallet(encrypted)
+          navigateToNextPage({
+            data: encrypted,
+            isWalletRestored: type === 'restore'
+          })
         })
         .catch(logger.error)
     }, 50) // allow UI render the spinner before async task
+
+    setSpinnerMessage(translate('screens/PinConfirmation', 'It may take a few seconds to secure and encrypt your wallet.'))
   }
 
-  async function discoverWalletAddresses (data: WalletPersistenceDataI<EncryptedProviderData>): Promise<void> {
-    const provider = await MnemonicEncrypted.initProvider(data, network, {
-      /**
-       * wallet context only use for READ purpose (non signing)
-       * see {@link TransactionAuthorization} for signing implementation
-       */
-      async prompt () {
-        throw new Error('No UI attached for passphrase prompting')
-      }
+  function navigateToNextPage (params: {data: WalletPersistenceDataI<EncryptedProviderData>, isWalletRestored: boolean}): void {
+    navigation.navigate({
+      name: 'WalletCreateRestoreSuccess', params, merge: true
     })
-    const wallet = await initJellyfishWallet(provider, network, client)
-
-    // get discovered address
-    const activeAddress = await wallet.discover(MAX_ALLOWED_ADDRESSES)
-
-    // sub 1 from total discovered address to get address index of last active address
-    const lastDiscoveredAddressIndex = Math.max(0, activeAddress.length - 1)
-    await WalletAddressIndexPersistence.setLength(lastDiscoveredAddressIndex)
   }
 
   return (
-    <ThemedScrollView
-      dark={tailwind('bg-gray-900')}
-      light={tailwind('bg-white')}
-      style={tailwind('w-full flex-1 flex-col')}
+    <ThemedScrollViewV2
+      style={tailwind('flex-1')}
+      contentContainerStyle={tailwind('pt-12 px-5 pb-16')}
+      testID='screen_confirm_pin'
     >
-      <CreateWalletStepIndicator
-        current={type === 'create' ? 3 : 2}
-        steps={type === 'create' ? CREATE_STEPS : RESTORE_STEPS}
-        style={tailwind('py-4 px-1')}
-      />
+      <View style={tailwind(['px-5 mb-5', { 'mb-12': spinnerMessage === undefined }])}>
+        <CreateWalletStepIndicatorV2
+          current={isCreateWallet ? 3 : 2}
+          steps={isCreateWallet ? CREATE_STEPS : RESTORE_STEPS}
+          style={tailwind(isCreateWallet ? 'px-4' : 'px-16')}
+          isComplete={isComplete}
+        />
 
-      <View style={tailwind('px-6 py-4 mb-6')}>
-        <ThemedText
-          style={tailwind('text-center font-semibold')}
+        <ThemedTextV2
+          style={tailwind('text-center font-normal-v2 mt-7')}
         >
-          {translate('screens/PinConfirmation', 'Enter your passcode again to verify')}
-        </ThemedText>
-      </View>
-
-      <PinTextInput
-        cellCount={6}
-        onChange={(pin) => {
-          setNewPin(pin)
-          verifyPin(pin)
-        }}
-        testID='pin_confirm_input'
-        value={newPin}
-      />
-
-      <View style={tailwind('flex-row justify-center mt-6')}>
+          {translate('screens/PinCreation', 'Add an additional layer of security by setting a passcode.')}
+        </ThemedTextV2>
         {
-          (spinnerMessage !== undefined)
-            ? (
-              <View style={tailwind('items-center px-4')}>
-                <ThemedActivityIndicator style={tailwind('mb-4')} />
-
-                <ThemedText
-                  style={tailwind('ml-2 font-semibold text-sm text-center w-4/6 px-4')}
-                >
-                  {spinnerMessage}
-                </ThemedText>
-              </View>
-              )
-            : null
-        }
-
-        {
-          invalid && (
-            <ThemedText
-              dark={tailwind('text-darkerror-500')}
-              light={tailwind('text-error-500')}
-              style={tailwind('text-center font-semibold text-sm')}
-              testID='wrong_passcode_text'
-            >
-              {translate('screens/PinConfirmation', 'Wrong passcode entered')}
-            </ThemedText>
+          (spinnerMessage !== undefined) && (
+            <ThemedActivityIndicatorV2 style={tailwind('py-4 my-0.5')} />
           )
         }
       </View>
-    </ThemedScrollView>
+
+      <View style={tailwind('mx-16', { 'mt-7': spinnerMessage === undefined })}>
+        <PinTextInputV2
+          cellCount={6}
+          onChange={(pin) => {
+            setNewPin(pin)
+            verifyPin(pin)
+          }}
+          testID='pin_confirm_input'
+          value={newPin}
+        />
+      </View>
+
+      <View style={tailwind('mt-5 px-7')}>
+        {
+          (spinnerMessage !== undefined) && (
+            <ThemedTextV2
+              style={tailwind('font-normal-v2 text-sm text-center')}
+              light={tailwind('text-mono-light-v2-700')}
+              dark={tailwind('text-mono-dark-v2-700')}
+            >
+              {spinnerMessage}
+            </ThemedTextV2>
+          )
+        }
+        {
+          (spinnerMessage === undefined && !invalid) && (
+            (
+              <ThemedTextV2
+                style={tailwind('text-sm font-normal-v2 text-center')}
+                light={tailwind('text-mono-light-v2-700')}
+                dark={tailwind('text-mono-dark-v2-700')}
+              >
+                {translate('screens/PinConfirmation', 'Enter passcode for verification')}
+              </ThemedTextV2>
+            )
+          )
+        }
+        {
+          invalid && (
+            <ThemedTextV2
+              style={tailwind('text-center font-normal-v2 text-sm text-red-v2')}
+              light={tailwind('text-red-v2')}
+              dark={tailwind('text-red-v2')}
+              testID='wrong_passcode_text'
+            >
+              {translate('screens/PinConfirmation', 'Wrong passcode entered')}
+            </ThemedTextV2>
+          )
+        }
+      </View>
+    </ThemedScrollViewV2>
   )
 }
