+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/PinCreation.tsx
@@ -2,12 +2,13 @@ import { NavigationProp, useNavigation } from '@react-navigation/native'
 import { StackScreenProps } from '@react-navigation/stack'
 import { useEffect, useState } from 'react'
 import { View } from '@components/index'
-import { CREATE_STEPS, CreateWalletStepIndicator, RESTORE_STEPS } from '@components/CreateWalletStepIndicator'
-import { PinTextInput } from '@components/PinTextInput'
-import { ThemedScrollView, ThemedText, ThemedTouchableOpacity, ThemedIcon } from '@components/themed'
+import { PinTextInputV2 } from '@components/PinTextInputV2'
+import { ThemedScrollViewV2, ThemedTextV2 } from '@components/themed'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import { WalletParamList } from '../../WalletNavigator'
+import { CREATE_STEPS, RESTORE_STEPS, CreateWalletStepIndicatorV2 } from '@components/CreateWalletStepIndicatorV2'
+import { LearnMoreCTA } from '../components/LearnModeCTA'
 
 type Props = StackScreenProps<WalletParamList, 'PinCreation'>
 
@@ -19,6 +20,7 @@ export function PinCreation ({ route }: Props): JSX.Element {
     type
   } = route.params
   const [newPin, setNewPin] = useState('')
+  const isCreateWallet = type === 'create'
   const goToPasscodeFaq = (): void => {
     navigation.navigate('PasscodeFaq')
   }
@@ -35,66 +37,43 @@ export function PinCreation ({ route }: Props): JSX.Element {
   }, [newPin])
 
   return (
-    <ThemedScrollView
-      dark={tailwind('bg-gray-900')}
-      light={tailwind('bg-white')}
-      style={tailwind('w-full flex-1 flex-col')}
-      contentContainerStyle={tailwind('items-center')}
+    <ThemedScrollViewV2
+      style={tailwind('flex-1')}
+      contentContainerStyle={tailwind('pt-12 px-5 pb-16')}
       testID='screen_create_pin'
     >
-      <CreateWalletStepIndicator
-        current={type === 'create' ? 3 : 2}
-        steps={type === 'create' ? CREATE_STEPS : RESTORE_STEPS}
-        style={tailwind('py-4 px-1')}
-      />
-
-      <View style={tailwind('px-6 py-4 mb-12')}>
-        <ThemedText
-          style={tailwind('text-center font-semibold')}
-        >
-          {translate('screens/PinCreation', `Well done! Your wallet is ${type === 'create' ? 'created' : 'restored'}.`)}
-        </ThemedText>
-      </View>
-
-      <PinTextInput
-        cellCount={6}
-        onChange={setNewPin}
-        testID='pin_input'
-        value={newPin}
-      />
-
-      <ThemedTouchableOpacity
-        onPress={goToPasscodeFaq}
-        light={tailwind('border-0')}
-        dark={tailwind('border-0')}
-        style={tailwind('w-4/5 pt-6')}
-        testID='passcode_faq_link'
-      >
-        <ThemedText
-          style={tailwind('text-center text-sm font-semibold')}
-          light={tailwind('text-primary-500')}
-          dark={tailwind('text-darkprimary-500')}
-        >
-          {translate('screens/PinCreation', 'Learn more about passcode')}
-        </ThemedText>
-      </ThemedTouchableOpacity>
-      <View style={tailwind('pt-12 flex flex-row text-sm font-medium rounded items-center')}>
-        <ThemedIcon
-          light={tailwind('text-gray-600')}
-          dark={tailwind('text-gray-400')}
-          style={tailwind('pr-1')}
-          iconType='MaterialIcons'
-          name='lock'
-          size={15}
+      <View style={tailwind('px-5 mb-12')}>
+        <CreateWalletStepIndicatorV2
+          current={isCreateWallet ? 3 : 2}
+          steps={isCreateWallet ? CREATE_STEPS : RESTORE_STEPS}
+          style={tailwind(isCreateWallet ? 'px-4' : 'px-16')}
         />
-        <ThemedText
-          light={tailwind('text-gray-500')}
-          dark={tailwind('text-gray-300')}
-          style={tailwind('text-xs')}
+        <ThemedTextV2
+          style={tailwind('text-center mt-7 text-center font-normal-v2')}
         >
-          {translate('screens/RecoveryWordsScreen', 'Keep your recovery words safe and private.')}
-        </ThemedText>
+          {translate('screens/PinCreation', 'Add an additional layer of security by setting a passcode.')}
+        </ThemedTextV2>
+        <LearnMoreCTA onPress={goToPasscodeFaq} testId='passcode_faq_link' />
       </View>
-    </ThemedScrollView>
+
+      <View style={tailwind('mx-16')}>
+        <PinTextInputV2
+          cellCount={6}
+          onChange={setNewPin}
+          testID='pin_input'
+          value={newPin}
+        />
+      </View>
+
+      <View style={tailwind('mt-5')}>
+        <ThemedTextV2
+          style={tailwind('text-sm font-normal-v2 text-center')}
+          light={tailwind('text-mono-light-v2-700')}
+          dark={tailwind('text-mono-dark-v2-700')}
+        >
+          {translate('screens/PinCreation', 'Create your passcode')}
+        </ThemedTextV2>
+      </View>
+    </ThemedScrollViewV2>
   )
 }
