+++ mobile-app/app/screens/WalletNavigator/screens/components/OnboardingCarousel.tsx
@@ -1,46 +1,54 @@
 import { Dimensions, Image, ImageSourcePropType, Platform } from 'react-native'
 import SwiperFlatList from 'react-native-swiper-flatlist'
-import ImageA from '@assets/images/onboarding/a.png'
-import ImageB from '@assets/images/onboarding/b.png'
-import ImageC from '@assets/images/onboarding/c.png'
-import ImageD from '@assets/images/onboarding/d.png'
+import ImageADark from '@assets/images/onboarding/welcome-screen-a-dark.png'
+import ImageBDark from '@assets/images/onboarding/welcome-screen-b-dark.png'
+import ImageCDark from '@assets/images/onboarding/welcome-screen-c-dark.png'
+import ImageDDark from '@assets/images/onboarding/welcome-screen-d-dark.png'
+import ImageALight from '@assets/images/onboarding/welcome-screen-a-light.png'
+import ImageBLight from '@assets/images/onboarding/welcome-screen-b-light.png'
+import ImageCLight from '@assets/images/onboarding/welcome-screen-c-light.png'
+import ImageDLight from '@assets/images/onboarding/welcome-screen-d-light.png'
 import { View } from '@components/index'
-import { AppIcon } from '@components/icons/AppIcon'
-import { ThemedText } from '@components/themed'
+import { ThemedTextV2 } from '@components/themed'
 import { useThemeContext } from '@shared-contexts/ThemeProvider'
-import { tailwind } from '@tailwind'
+import { getColor, tailwind } from '@tailwind'
 import { translate } from '@translations'
-import { VersionTag } from '@components/VersionTag'
+import { CarouselPagination } from '@screens/WalletNavigator/screens/components/CarouselPagination'
 
 interface CarouselImage {
-  image: ImageSourcePropType
+  imageDark: ImageSourcePropType
+  imageLight: ImageSourcePropType
   title: string
-  secondTitle?: string
   subtitle: string
 }
 
-const slides: JSX.Element[] = [<InitialSlide key={0} />,
+const slides: JSX.Element[] = [
   <ImageSlide
-    image={ImageA}
-    key={1}
-    subtitle='DeFiChain Wallet is fully non-custodial. Keep your 24-word recovery phrase safe. Only you have access to your funds.'
+    imageDark={ImageADark}
+    imageLight={ImageALight}
+    key={0}
+    subtitle='DeFiChain Wallet is fully non-custodial. Only you will have access to your fund.'
     title='Take full control'
   />,
   <ImageSlide
-    image={ImageB}
-    key={2}
+    imageDark={ImageBDark}
+    imageLight={ImageBLight}
+    key={1}
     subtitle='Review your available and locked assets in your portfolio.'
     title='View your assets in one place'
   />,
   <ImageSlide
-    image={ImageC}
-    key={3}
+    imageDark={ImageCDark}
+    imageLight={ImageCLight}
+    key={2}
     subtitle='Trade on the DEX and earn rewards from liquidity mining with crypto and dTokens.'
     title='Maximize earning potential'
   />,
+
   <ImageSlide
-    image={ImageD}
-    key={4}
+    imageDark={ImageDDark}
+    imageLight={ImageDLight}
+    key={3}
     subtitle='Access financial opportunities with dTokens minted through decentralized vaults.'
     title='Decentralized loans'
   />]
@@ -48,60 +56,26 @@ const slides: JSX.Element[] = [<InitialSlide key={0} />,
 // Needs for it to work on web. Otherwise, it takes full window size
 const { width } = Platform.OS === 'web' ? { width: '375px' } : Dimensions.get('window')
 
-export function InitialSlide (): JSX.Element {
+export function ImageSlide ({ imageDark, imageLight, title, subtitle }: CarouselImage): JSX.Element {
+  const { isLight } = useThemeContext()
   return (
-    <View style={tailwind('flex-1 items-center justify-center p-8')}>
-      <AppIcon
-        height={100}
-        width={100}
+    <View style={tailwind('flex-1 items-center justify-center px-10')}>
+      <Image
+        source={isLight ? imageLight : imageDark}
+        style={{ width: 204, height: 136 }}
       />
-
-      <ThemedText style={tailwind('text-2xl font-bold mt-3')}>
-        {translate('screens/OnboardingCarousel', 'DeFiChain Wallet')}
-      </ThemedText>
-
-      <ThemedText
-        dark={tailwind('text-gray-400')}
-        light={tailwind('text-gray-500')}
-        style={tailwind('text-base font-medium mt-1')}
-      >
-        {translate('screens/OnboardingCarousel', 'Native DeFi for Bitcoin')}
-      </ThemedText>
-
-      <View style={tailwind('mt-2')}>
-        <VersionTag />
-      </View>
-    </View>
-  )
-}
-
-export function ImageSlide ({ image, title, secondTitle, subtitle }: CarouselImage): JSX.Element {
-  return (
-    <View style={tailwind('flex-1 items-center justify-center py-8 px-5')}>
-      <View style={tailwind('h-2/6 items-center justify-center')}>
-        <ThemedText style={tailwind('text-2xl font-bold text-center')}>
+      <View style={tailwind('items-center justify-center mt-7 mb-5')}>
+        <ThemedTextV2
+          style={tailwind('text-xl font-semibold-v2 text-center')}
+        >
           {translate('screens/OnboardingCarousel', title)}
-        </ThemedText>
-
-        {secondTitle !== undefined && (
-          <ThemedText style={tailwind('text-2xl font-bold text-center')}>
-            {translate('screens/OnboardingCarousel', secondTitle)}
-          </ThemedText>
-        )}
-
-        <ThemedText
-          dark={tailwind('text-gray-400')}
-          light={tailwind('text-gray-500')}
-          style={tailwind('font-normal text-center mt-1 mb-8')}
+        </ThemedTextV2>
+        <ThemedTextV2
+          style={tailwind('font-normal-v2 text-center mt-2')}
         >
           {translate('screens/OnboardingCarousel', subtitle)}
-        </ThemedText>
+        </ThemedTextV2>
       </View>
-
-      <Image
-        source={image}
-        style={{ width, height: '55%' }}
-      />
     </View>
   )
 }
@@ -116,9 +90,11 @@ export function OnboardingCarousel (): JSX.Element {
       autoplayLoopKeepAnimation
       data={slides}
       index={0}
-      paginationActiveColor={isLight ? 'rgba(0, 0, 0, 0.8)' : '#D4D4D4'}
-      paginationDefaultColor={isLight ? 'rgba(0, 0, 0, 0.1)' : '#262626'}
-      paginationStyleItem={tailwind('h-2.5 w-2.5 mx-1.5')}
+      paginationActiveColor={isLight ? getColor('mono-light-v2-900') : getColor('mono-dark-v2-900')}
+      paginationStyleItemActive={tailwind('w-6 h-1.5')}
+      paginationDefaultColor={isLight ? getColor('mono-light-v2-500') : getColor('mono-dark-v2-500')}
+      paginationStyleItem={tailwind('h-1.5 w-1.5 mx-0.75')}
+      PaginationComponent={CarouselPagination}
       renderItem={({ item }) => (
         <View style={{ width }}>
           {
